Final Project/zzzz/requirements.txt
Flask==2.0.1
pymongo==4.1.1
geopy==2.2.0
python-dotenv==0.19.0
Werkzeug==2.0.1
Flask-Limiter==3.5.0
gunicorn==21.2.0
pytest==7.4.2
black==23.7.0
flake8==6.1.0
python-jose[cryptography]==3.3.0
requests==2.26.0
PyJWT==2.8.0
cryptography==41.0.7
fast2sms==1.0.0
pywhatkit==5.4.0
translate==3.6.1
google-generativeai==0.3.1 
pytz
fastapi==0.115.0
uvicorn==0.30.6
twilio==9.3.3
# Celery and Redis for Agentic AI
celery==5.3.4
redis==4.6.0
celery[redis]==5.3.4
# Additional dependencies for agents
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
